Add live bundler for trails-config
Add live bundle for trails-config. The trails serve command can watch
a directory with a folder "config" that contains all the config
according to the open-osquery config namespacing and the config will be
served as a static bundle on an http endpoint for the trails extension
to consume.

The config files are bundled and signed with every change atmost once
per second.

TODO: unittests to follow in a following PR.